requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
asgiref==3.9.2
    # via
    #   -r requirements/base.txt
    #   django
click==8.3.0
    # via code-annotations
code-annotations==2.3.0
    # via -r requirements/test.in
coverage[toml]==7.10.7
    # via pytest-cov
django==4.2.24
    # via
    #   -c https:/raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   djangorestframework
djangorestframework==3.16.1
    # via -r requirements/base.txt
factory-boy==3.3.3
    # via -r requirements/test.in
faker==37.8.0
    # via factory-boy
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via code-annotations
markupsafe==3.0.2
    # via jinja2
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pygments==2.19.2
    # via pytest
pytest==8.4.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==7.0.0
    # via -r requirements/test.in
pytest-django==4.11.1
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via code-annotations
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.5.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tzdata==2025.2
    # via faker
